fix(tokens): bind estimates to selected identity
Carry the validated capability token identity with each pinned request pricing snapshot. Exact OpenAI identities now fail closed when no tokenizer exists, while explicitly approximate provider identities cannot become exact through model-name inference. Apply the same contract to cache hits and embedding reservations.

Fixes #1208
