+++ src/utils/ai/counter/token_counter.rs
@@ -2,8 +2,6 @@
 
 use super::types::{ModelTokenConfig, TokenEstimate};
 use crate::core::models::openai::{ChatMessage, ContentPart, MessageContent};
-use crate::core::providers::openai::models::get_openai_registry;
-use crate::core::types::model_id::ModelIdRef;
 use crate::utils::error::gateway_error::{GatewayError, Result};
 use std::collections::HashMap;
 use tiktoken_rs::{ChatCompletionRequestMessage, bpe_for_model, num_tokens_from_messages};
@@ -15,11 +13,84 @@ pub struct TokenCounter {
     model_configs: HashMap<String, ModelTokenConfig>,
 }
 
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub(crate) enum TokenizerIdentity {
+    ExactOpenAi(String),
+    Approximate { provider: String, model: String },
+}
+
+impl TokenizerIdentity {
+    pub(crate) fn exact_openai(model: impl Into<String>) -> Self {
+        Self::ExactOpenAi(model.into())
+    }
+
+    pub(crate) fn approximate(provider: impl Into<String>, model: impl Into<String>) -> Self {
+        Self::Approximate {
+            provider: provider.into(),
+            model: model.into(),
+        }
+    }
+
+    pub(crate) fn provider(&self) -> &str {
+        match self {
+            Self::ExactOpenAi(_) => "openai",
+            Self::Approximate { provider, .. } => provider,
+        }
+    }
+
+    pub(crate) fn model(&self) -> &str {
+        match self {
+            Self::ExactOpenAi(model) | Self::Approximate { model, .. } => model,
+        }
+    }
+
+    fn tokenizer_model(&self) -> Result<&str> {
+        if matches!(self, Self::ExactOpenAi(_)) && exact_bpe_for_model(self.model()).is_err() {
+            return Err(GatewayError::Config(format!(
+                "tokenizer unavailable for exact identity '{}/{}'",
+                self.provider(),
+                self.model()
+            )));
+        }
+        Ok(self.model())
+    }
+}
+
 impl TokenCounter {
     /// Create a new token counter
     pub fn new() -> Self {
         Self {
             model_configs: ModelTokenConfig::default_configs(),
+        }
+    }
+
+    pub(crate) fn count_completion_tokens_for_identity(
+        &self,
+        identity: &TokenizerIdentity,
+        prompt: &str,
+    ) -> Result<TokenEstimate> {
+        match identity {
+            TokenizerIdentity::ExactOpenAi(_) => {
+                self.count_completion_tokens(identity.tokenizer_model()?, prompt)
+            }
+            TokenizerIdentity::Approximate { model, .. } => {
+                self.approximate_completion_tokens(model, prompt)
+            }
+        }
+    }
+
+    pub(crate) fn count_chat_tokens_for_identity(
+        &self,
+        identity: &TokenizerIdentity,
+        messages: &[ChatMessage],
+    ) -> Result<TokenEstimate> {
+        match identity {
+            TokenizerIdentity::ExactOpenAi(_) => {
+                self.count_chat_tokens(identity.tokenizer_model()?, messages)
+            }
+            TokenizerIdentity::Approximate { model, .. } => {
+                self.approximate_chat_tokens(model, messages)
+            }
         }
     }
 
@@ -29,7 +100,6 @@ impl TokenCounter {
         model: &str,
         messages: &[ChatMessage],
     ) -> Result<TokenEstimate> {
-        ensure_explicit_openai_tokenizer(model)?;
         if let Some(total_tokens) = self.exact_chat_tokens(model, messages)? {
             return Ok(TokenEstimate {
                 input_tokens: total_tokens,
@@ -40,6 +110,14 @@ impl TokenCounter {
             });
         }
 
+        self.approximate_chat_tokens(model, messages)
+    }
+
+    fn approximate_chat_tokens(
+        &self,
+        model: &str,
+        messages: &[ChatMessage],
+    ) -> Result<TokenEstimate> {
         let config = self.get_model_config(model)?;
         let mut total_tokens = config.request_overhead;
 
@@ -155,7 +233,6 @@ impl TokenCounter {
 
     /// Count tokens in completion request
     pub fn count_completion_tokens(&self, model: &str, prompt: &str) -> Result<TokenEstimate> {
-        ensure_explicit_openai_tokenizer(model)?;
         if let Some(input_tokens) = exact_text_tokens(model, prompt)? {
             return Ok(TokenEstimate {
                 input_tokens,
@@ -166,6 +243,10 @@ impl TokenCounter {
             });
         }
 
+        self.approximate_completion_tokens(model, prompt)
+    }
+
+    fn approximate_completion_tokens(&self, model: &str, prompt: &str) -> Result<TokenEstimate> {
         let config = self.get_model_config(model)?;
         let input_tokens = config.request_overhead + self.estimate_text_tokens(config, prompt);
 
@@ -180,7 +261,6 @@ impl TokenCounter {
 
     /// Count tokens in embedding request
     pub fn count_embedding_tokens(&self, model: &str, input: &[String]) -> Result<TokenEstimate> {
-        ensure_explicit_openai_tokenizer(model)?;
         let exact_counts = input
             .iter()
             .map(|text| exact_text_tokens(model, text))
@@ -219,7 +299,6 @@ impl TokenCounter {
         input_tokens: u32,
         model: &str,
     ) -> Result<u32> {
-        ensure_explicit_openai_tokenizer(model)?;
         let config = self.get_model_config(model)?;
 
         if let Some(max) = max_tokens {
@@ -240,7 +319,6 @@ impl TokenCounter {
         input_tokens: u32,
         max_output_tokens: Option<u32>,
     ) -> Result<bool> {
-        ensure_explicit_openai_tokenizer(model)?;
         let config = self.get_model_config(model)?;
         let output_tokens = max_output_tokens.unwrap_or(0);
         let total_tokens = input_tokens + output_tokens;
@@ -363,41 +441,6 @@ fn exact_bpe_for_model(model: &str) -> std::result::Result<&'static tiktoken_rs:
     bpe_for_model(tokenizer_model_name(model)).map_err(|_| ())
 }
 
-fn explicit_openai_catalog_model(model: &str) -> Option<&str> {
-    let parsed = ModelIdRef::parse(model);
-    parsed
-        .provider()
-        .is_some_and(|provider| provider.eq_ignore_ascii_case("openai"))
-        .then(|| parsed.model())
-}
-
-fn unknown_openai_catalog_model_error(model: &str) -> GatewayError {
-    GatewayError::Config(format!("unknown OpenAI catalog model '{model}'"))
-}
-
-fn openai_tokenizer_unavailable_error(model: &str) -> GatewayError {
-    GatewayError::Config(format!(
-        "tokenizer unavailable for OpenAI catalog model '{model}'"
-    ))
-}
-
-fn ensure_explicit_openai_tokenizer(model: &str) -> Result<()> {
-    let Some(catalog_model) = explicit_openai_catalog_model(model) else {
-        return Ok(());
-    };
-    if get_openai_registry()
-        .get_model_spec(catalog_model)
-        .is_none()
-    {
-        return Err(unknown_openai_catalog_model_error(model));
-    }
-    if exact_bpe_for_model(catalog_model).is_err() {
-        return Err(openai_tokenizer_unavailable_error(model));
-    }
-
-    Ok(())
-}
-
 fn plain_text_message_content(message: &ChatMessage) -> Option<Option<String>> {
     match &message.content {
         None => Some(None),
@@ -418,379 +461,6 @@ fn usize_to_u32(value: usize) -> Result<u32> {
         .map_err(|_| GatewayError::Config(format!("token count {value} exceeds u32 range")))
 }
 
-// ==================== Unit Tests ====================
-
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    // ==================== TokenCounter Creation Tests ====================
-
-    #[test]
-    fn test_token_counter_new() {
-        let counter = TokenCounter::new();
-        assert!(!counter.model_configs.is_empty());
-    }
-
-    #[test]
-    fn test_token_counter_default() {
-        let counter = TokenCounter::default();
-        assert!(!counter.model_configs.is_empty());
-    }
-
-    #[test]
-    fn test_token_counter_clone() {
-        let counter = TokenCounter::new();
-        let cloned = counter.clone();
-        assert_eq!(counter.model_configs.len(), cloned.model_configs.len());
-    }
-
-    #[test]
-    fn test_token_counter_debug() {
-        let counter = TokenCounter::new();
-        let debug_str = format!("{:?}", counter);
-        assert!(debug_str.contains("TokenCounter"));
-        assert!(debug_str.contains("model_configs"));
-    }
-
-    // ==================== Model Config Tests ====================
-
-    #[test]
-    fn test_get_model_config_exact_match() {
-        let counter = TokenCounter::new();
-        let config = counter.get_model_config("gpt-4");
-        assert!(config.is_ok());
-    }
-
-    #[test]
-    fn test_get_model_config_gpt4_variant() {
-        let counter = TokenCounter::new();
-        let config = counter.get_model_config("gpt-4-turbo");
-        assert!(config.is_ok());
-    }
-
-    #[test]
-    fn test_get_model_config_gpt35() {
-        let counter = TokenCounter::new();
-        let config = counter.get_model_config("gpt-3.5-turbo");
-        assert!(config.is_ok());
-    }
-
-    #[test]
-    fn test_get_model_config_claude() {
-        let counter = TokenCounter::new();
-        let config = counter.get_model_config("claude-3-opus");
-        assert!(config.is_ok());
-    }
-
-    #[test]
-    fn test_get_model_config_fallback_to_default() {
-        let counter = TokenCounter::new();
-        let config = counter.get_model_config("unknown-model-xyz");
-        // Should fall back to default config
-        assert!(config.is_ok());
-    }
-
-    // ==================== Model Family Extraction Tests ====================
-
-    #[test]
-    fn test_extract_model_family_gpt4() {
-        let counter = TokenCounter::new();
-        assert_eq!(counter.extract_model_family("gpt-4"), "gpt-4");
-        assert_eq!(counter.extract_model_family("gpt-4-turbo"), "gpt-4");
-        assert_eq!(counter.extract_model_family("gpt-4-0125-preview"), "gpt-4");
-        assert_eq!(counter.extract_model_family("gpt-4o"), "gpt-4");
-    }
-
-    #[test]
-    fn test_extract_model_family_gpt35() {
-        let counter = TokenCounter::new();
-        assert_eq!(
-            counter.extract_model_family("gpt-3.5-turbo"),
-            "gpt-3.5-turbo"
-        );
-        assert_eq!(
-            counter.extract_model_family("gpt-3.5-turbo-16k"),
-            "gpt-3.5-turbo"
-        );
-    }
-
-    #[test]
-    fn test_extract_model_family_claude() {
-        let counter = TokenCounter::new();
-        assert_eq!(counter.extract_model_family("claude-3-opus"), "claude-3");
-        assert_eq!(counter.extract_model_family("claude-3-sonnet"), "claude-3");
-        assert_eq!(counter.extract_model_family("claude-3-haiku"), "claude-3");
-        assert_eq!(counter.extract_model_family("claude-2.1"), "claude-2");
-    }
-
-    #[test]
-    fn test_extract_model_family_with_provider_prefix() {
-        let counter = TokenCounter::new();
-        assert_eq!(counter.extract_model_family("openai/gpt-4"), "gpt-4");
-        assert_eq!(
-            counter.extract_model_family("anthropic/claude-3-opus"),
-            "claude-3"
-        );
-    }
-
-    #[test]
-    fn test_extract_model_family_unknown() {
-        let counter = TokenCounter::new();
-        assert_eq!(counter.extract_model_family("unknown-model"), "default");
-        assert_eq!(counter.extract_model_family("llama-2-70b"), "default");
-    }
-
-    // ==================== Text Token Estimation Tests ====================
-
-    #[test]
-    fn test_estimate_text_tokens_empty() {
-        let counter = TokenCounter::new();
-        let config = counter.get_model_config("gpt-4").unwrap();
-        let tokens = counter.estimate_text_tokens(config, "");
-        assert_eq!(tokens, 0);
-    }
-
-    #[test]
-    fn test_estimate_text_tokens_short_text() {
-        let counter = TokenCounter::new();
-        let config = counter.get_model_config("gpt-4").unwrap();
-        let tokens = counter.estimate_text_tokens(config, "Hello");
-        assert!(tokens > 0);
-        assert!(tokens < 10);
-    }
-
-    #[test]
-    fn test_estimate_text_tokens_longer_text() {
-        let counter = TokenCounter::new();
-        let config = counter.get_model_config("gpt-4").unwrap();
-        let short_tokens = counter.estimate_text_tokens(config, "Hello");
-        let long_tokens =
-            counter.estimate_text_tokens(config, "Hello, this is a much longer text message.");
-        assert!(long_tokens > short_tokens);
-    }
-
-    #[test]
-    fn test_estimate_text_tokens_unicode() {
-        let counter = TokenCounter::new();
-        let config = counter.get_model_config("gpt-4").unwrap();
-        let tokens = counter.estimate_text_tokens(config, "你好世界");
-        assert!(tokens > 0);
-    }
-
-    // ==================== Completion Token Counting Tests ====================
-
-    #[test]
-    fn test_count_completion_tokens_basic() {
-        let counter = TokenCounter::new();
-        let result = counter.count_completion_tokens("gpt-4", "Hello, world!");
-        assert!(result.is_ok());
-        let estimate = result.unwrap();
-        assert!(estimate.input_tokens > 0);
-        assert!(!estimate.is_approximate);
-    }
-
-    #[test]
-    fn test_count_completion_tokens_empty() {
-        let counter = TokenCounter::new();
-        let result = counter.count_completion_tokens("gpt-4", "");
-        assert!(result.is_ok());
-        let estimate = result.unwrap();
-        assert_eq!(estimate.input_tokens, 0);
-        assert!(!estimate.is_approximate);
-    }
-
-    #[test]
-    fn test_count_completion_tokens_long_text() {
-        let counter = TokenCounter::new();
-        let long_text = "word ".repeat(1000);
-        let result = counter.count_completion_tokens("gpt-4", &long_text);
-        assert!(result.is_ok());
-        let estimate = result.unwrap();
-        assert!(estimate.input_tokens > 100);
-    }
-
-    #[test]
-    fn test_count_completion_tokens_confidence() {
-        let counter = TokenCounter::new();
-        let result = counter.count_completion_tokens("gpt-4", "test");
-        assert!(result.is_ok());
-        let estimate = result.unwrap();
-        assert!(estimate.confidence > 0.0);
-        assert!(estimate.confidence <= 1.0);
-    }
-
-    // ==================== Token Estimate Tests ====================
-
-    #[test]
-    fn test_token_estimate_structure() {
-        let counter = TokenCounter::new();
-        let result = counter.count_completion_tokens("gpt-4", "Hello");
-        assert!(result.is_ok());
-        let estimate = result.unwrap();
-
-        assert_eq!(estimate.total_tokens, estimate.input_tokens);
-        assert!(estimate.output_tokens.is_none());
-        assert!(!estimate.is_approximate);
-    }
-
-    // ==================== Embedding Token Counting Tests ====================
-
-    #[test]
-    fn test_count_embedding_tokens_single() {
-        let counter = TokenCounter::new();
-        let input = vec!["Hello, world!".to_string()];
-        let result = counter.count_embedding_tokens("gpt-4", &input);
-        assert!(result.is_ok());
-        let estimate = result.unwrap();
-        assert!(estimate.input_tokens > 0);
-        assert_eq!(estimate.confidence, 1.0);
-        assert!(!estimate.is_approximate);
-    }
-
-    #[test]
-    fn test_count_embedding_tokens_multiple() {
-        let counter = TokenCounter::new();
-        let input = vec![
-            "First text".to_string(),
-            "Second text".to_string(),
-            "Third text".to_string(),
-        ];
-        let result = counter.count_embedding_tokens("gpt-4", &input);
-        assert!(result.is_ok());
-        let estimate = result.unwrap();
-        assert!(estimate.input_tokens > 0);
-    }
-
-    #[test]
-    fn test_count_embedding_tokens_empty() {
-        let counter = TokenCounter::new();
-        let input: Vec<String> = vec![];
-        let result = counter.count_embedding_tokens("gpt-4", &input);
-        assert!(result.is_ok());
-    }
-
-    // ==================== Output Token Estimation Tests ====================
-
-    #[test]
-    fn test_estimate_output_tokens_with_max() {
-        let counter = TokenCounter::new();
-        let result = counter.estimate_output_tokens(Some(100), 50, "gpt-4");
-        assert!(result.is_ok());
-        let output = result.unwrap();
-        assert_eq!(output, 100);
-    }
-
-    #[test]
-    fn test_estimate_output_tokens_without_max() {
-        let counter = TokenCounter::new();
-        let result = counter.estimate_output_tokens(None, 100, "gpt-4");
-        assert!(result.is_ok());
-        let output = result.unwrap();
-        // Should be ~25% of remaining context
-        assert!(output > 0);
-    }
-
-    #[test]
-    fn test_estimate_output_tokens_capped_by_context() {
-        let counter = TokenCounter::new();
-        // Request more tokens than available
-        let result = counter.estimate_output_tokens(Some(1_000_000), 0, "gpt-4");
-        assert!(result.is_ok());
-        let output = result.unwrap();
-        // Should be capped at model's max context
-        assert!(output < 1_000_000);
-    }
-
-    // ==================== Context Window Check Tests ====================
-
-    #[test]
-    fn test_check_context_window_fits() {
-        let counter = TokenCounter::new();
-        let result = counter.check_context_window("gpt-4", 1000, Some(1000));
-        assert!(result.is_ok());
-        assert!(result.unwrap());
-    }
-
-    #[test]
-    fn test_check_context_window_exceeds() {
-        let counter = TokenCounter::new();
-        // Try to use more tokens than the context window
-        let result = counter.check_context_window("gpt-4", 100_000, Some(100_000));
-        assert!(result.is_ok());
-        // Verify the function returns a boolean (regardless of value)
-        let _fits = result.unwrap();
-    }
-
-    #[test]
-    fn test_check_context_window_no_output() {
-        let counter = TokenCounter::new();
-        let result = counter.check_context_window("gpt-4", 1000, None);
-        assert!(result.is_ok());
-        assert!(result.unwrap());
-    }
-
-    // ==================== Model Config Management Tests ====================
-
-    #[test]
-    fn test_add_model_config() {
-        let mut counter = TokenCounter::new();
-        let initial_count = counter.model_configs.len();
-
-        let config = ModelTokenConfig {
-            model: "custom-model".to_string(),
-            chars_per_token: 4.5,
-            message_overhead: 5,
-            request_overhead: 10,
-            max_context_tokens: 16000,
-            special_tokens: HashMap::new(),
-        };
-        counter.add_model_config(config);
-
-        assert_eq!(counter.model_configs.len(), initial_count + 1);
-        assert!(counter.get_model_config("custom-model").is_ok());
-    }
-
-    #[test]
-    fn test_get_supported_models() {
-        let counter = TokenCounter::new();
-        let models = counter.get_supported_models();
-        assert!(!models.is_empty());
-    }
-
-    // ==================== Edge Cases Tests ====================
-
-    #[test]
-    fn test_special_characters() {
-        let counter = TokenCounter::new();
-        let config = counter.get_model_config("gpt-4").unwrap();
-        let tokens = counter.estimate_text_tokens(config, "!@#$%^&*()");
-        assert!(tokens > 0);
-    }
-
-    #[test]
-    fn test_newlines_and_whitespace() {
-        let counter = TokenCounter::new();
-        let config = counter.get_model_config("gpt-4").unwrap();
-        let tokens = counter.estimate_text_tokens(config, "Hello\n\n\nWorld\t\tTest");
-        assert!(tokens > 0);
-    }
-
-    #[test]
-    fn test_very_long_word() {
-        let counter = TokenCounter::new();
-        let config = counter.get_model_config("gpt-4").unwrap();
-        let long_word = "a".repeat(1000);
-        let tokens = counter.estimate_text_tokens(config, &long_word);
-        assert!(tokens > 0);
-    }
-
-    #[test]
-    fn test_mixed_content() {
-        let counter = TokenCounter::new();
-        let config = counter.get_model_config("gpt-4").unwrap();
-        let mixed = "Hello 你好 Привет مرحبا 🎉";
-        let tokens = counter.estimate_text_tokens(config, mixed);
-        assert!(tokens > 0);
-    }
-}
+#[path = "token_counter_tests.rs"]
+mod tests;
